Upgrade for the project - v2.0a
Two new columns were added to display in the data tables, the possibility of setting the entry date manually was also removed and the way the RAM capacity is recorded was changed to make it much easier.  And also fixed a bug that made it impossible to record data or modify it, and more records were added from the database side. (It remains to add the update of the DB)